remove double class in single file
