+++ packages/kbn-ui-actions-browser/src/service/ui_actions_execution_service.ts
@@ -10,17 +10,7 @@ import { uniqBy } from 'lodash';
 import { Action } from '../actions';
 import { buildContextMenuForActions, openContextMenu } from '../context_menu';
 import { Trigger } from '../triggers';
-
-export class Defer<T> {
-  public readonly resolve!: (data: T) => void;
-  public readonly reject!: (error: any) => void;
-  public readonly promise: Promise<T> = new Promise<T>((resolve, reject) => {
-    (this as any).resolve = resolve;
-    (this as any).reject = reject;
-  });
-}
-
-export const createDefer = <T>() => new Defer<T>();
+import { createDefer, type Defer } from './defer';
 
 interface ExecuteActionTask {
   action: Action;
